nhl-betting/deploy/requirements.txt
lightgbm>=4.0
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
psycopg2-binary>=2.9
requests>=2.31
pyyaml>=6.0
python-dotenv>=1.0
joblib>=1.3
schedule>=1.2
